GitRepository to GitRepositoryUtils
